requirements.txt
fastapi
uvicorn[standard]
pymongo
faiss-cpu
pandas
requests
numpy
